+++ server2/build.gradle
@@ -38,8 +38,6 @@ subprojects { // 각 모듈에 적용할 공통 설정
         testImplementation 'org.springframework.amqp:spring-rabbit-test'
         testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
         implementation 'org.springframework.boot:spring-boot-starter-data-mongodb'
-
-
     }
 
     tasks.named('bootJar') { //빌드할 때 bootjar 파일로 하겠다는 의미
