--- server2/server-child/src/main/java/sinhan/server1/Server2Application.java
@@ -1,18 +1,0 @@
-package sinhan.server1;
-
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Server2Application {
-
-    public static void main(String[] args) {
-        SpringApplication.run(Server3Application.class, args);
-    }
-
-//    @RabbitListener(queues = "hello")
-//    public void listen(String message) {
-//        System.out.println("Received: " + message);
-//    }
-}
