--- server2/server-child/src/main/java/sinhan/server1/domain/AccountController.java
@@ -1,13 +1,0 @@
-package sinhan.server1.domain;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class AccountController {
-    @GetMapping("/temp")
-    public String getStock(){
-        return "addff";
-    }
-
-}
