--- server2/server-common/src/main/java/sinhan/server1/domain/invest/entity/StockDivideOutput.java
@@ -1,15 +1,0 @@
-package sinhan.server1.domain.invest.entity;
-
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategy;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import java.util.List;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
-public class StockDivideOutput {
-    List<StockDivideDetail> list;
-}
