--- server2/server-common/src/main/java/sinhan/server1/domain/invest/repository/StockRepository.java
@@ -1,7 +1,0 @@
-package sinhan.server1.domain.invest.repository;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class StockRepository {
-}
