+++ server2/server-common/src/main/java/sinhan/server2/domain/invest/entity/StockCurrentPriceOutput.java
@@ -1,0 +1,15 @@
+package sinhan.server2.domain.invest.entity;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.PropertyNamingStrategy;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
+public class StockCurrentPriceOutput {
+    @JsonProperty("thstrm")
+    private String dividendRate;
+}
