+++ server2/server-common/src/main/java/sinhan/server2/domain/invest/entity/StockDuraionPriceOutput.java
@@ -1,0 +1,14 @@
+package sinhan.server2.domain.invest.entity;
+
+import java.util.List;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class StockDuraionPriceOutput {
+    
+    private StockDurationDetail output1;
+    private List<StockCartDate> output2;
+
+}
