+++ server2/server-common/src/main/java/sinhan/server2/domain/invest/entity/StockFianceResponseOutput.java
@@ -1,0 +1,13 @@
+package sinhan.server2.domain.invest.entity;
+
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import java.util.List;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@JsonNaming
+public class StockFianceResponseOutput {
+    public List<StockFianceDetail> output;
+}
