+++ server2/server-common/src/main/java/sinhan/server2/domain/invest/repository/StockRepository.java
@@ -1,0 +1,7 @@
+package sinhan.server2.domain.invest.repository;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class StockRepository {
+}
