+++ server2/server-common/src/main/java/sinhan/server2/domain/tempuser/TempUserRepository.java
@@ -1,0 +1,7 @@
+package sinhan.server2.domain.tempuser;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TempUserRepository extends JpaRepository<TempUser,Integer> {
+    TempUser findBySerialNumber(Long serialNumber);
+}
