+++ server2/server-common/src/main/java/sinhan/server2/global/exception/CustomException.java
@@ -1,0 +1,12 @@
+package sinhan.server2.global.exception;
+
+import lombok.Getter;
+
+@Getter
+public class CustomException extends RuntimeException {
+    private final ErrorCode errorCode;
+    public CustomException(ErrorCode errorCode) {
+        super(errorCode.getMessage());
+        this.errorCode = errorCode;
+    }
+}
