+++ server2/server-common/src/main/java/sinhan/server2/global/exception/CustomExceptionHandler.java
@@ -1,0 +1,31 @@
+package sinhan.server2.global.exception;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@Slf4j
+@RestControllerAdvice
+public class CustomExceptionHandler {
+
+    @ExceptionHandler
+    protected ResponseEntity<ErrorResponse> handleCustomException(CustomException e) {
+        return ErrorResponse.toResponseEntity(e);
+    }
+
+    // bean validation 관련 에러
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    protected ResponseEntity<ErrorResponse> handleMethodArgumentNotValidException(MethodArgumentNotValidException e) {
+
+        FieldError fieldError = e.getFieldError();
+
+        log.info(fieldError.toString());
+        return ErrorResponse.toResponseEntity(HttpStatus.BAD_REQUEST, fieldError);
+    }
+
+
+}
